requirement.txt
pyfiglet
cryptography
pyyaml